synchronous-agent-invocation/requirements.txt
requests==2.31.0
sensa-python==6.3.1
websocket-client==1.2.1
fastapi==0.68.1
uvicorn==0.15.0